+++ sh4asm-ts/CHANGELOG.md
@@ -13,3 +13,7 @@
 ## [0.0.7]
 * Fixed missing snippets.
 * Expanded regex for stage and char hover information.
+
+## [0.0.8]
+* Added text-color hover information by using `text` at the end of a line.
+* Added snippets for text-colors.